[Refactor] 페이지 요소 위치 변경
