+++ src/App.js
@@ -12,7 +12,8 @@ import KeyWord from './pages/Onboarding/KeyWord.js';
 import Intro from './pages/Onboarding/Intro';
 import MapMain from './pages/map/Main/MapMain';
 import MapSearch from './pages/map/search/MapSearch';
-
+import Home from'./pages/Main/Home'
+import LikeMain from './pages/Like/LikeMain.js'
 
 function App() {
   return (
