+++ src/pages/Main/Home.jsx
@@ -10,11 +10,11 @@ import { Navigation, Pagination, Autoplay } from 'swiper/modules';
 import { ReactComponent as Banner1 } from '../../assets/banner1.svg';
 import { ReactComponent as UpIcon } from '../../assets/up_icon.svg';
 // 컴포넌트
-import Header from '../../component/main/Header';
-import Navbar from '../../component/main/Navbar';
-import Recommend from '../../component/main/Recommend/Recommend';
-import Top5 from '../../component/main/Top5/Top5';
-import Community from '../../component/main/Community/Community';
+import Header from '../../component/Main/Header';
+import Navbar from '../../component/Main/Navbar';
+import Recommend from '../../component/Main/Recommend/Recommend';
+import Top5 from '../../component/Main/Top5/Top5';
+import Community from '../../component/Main/Community/Community';
 
 function Home() {
